+++ app/core/models/__init__.py
@@ -1,6 +1,7 @@
 from .audit import CenturionAudit    # pylint: disable=W0611:unused-import
 from . import ticket_comment_action    # pylint: disable=W0611:unused-import
 from . import ticket_comment_action_field_edit    # pylint: disable=W0611:unused-import
+from . import ticket_comment_action_model_link    # pylint: disable=W0611:unused-import
 from . import ticket_comment_base    # pylint: disable=W0611:unused-import
 from . import ticket_comment_solution    # pylint: disable=W0611:unused-import
 from . import ticket_comment_task    # pylint: disable=W0611:unused-import