+++ app/core/models/ticket_comment_action_model_link.py
@@ -1,0 +1,76 @@
+from django.contrib.auth.models import (
+    ContentType,
+)
+from django.db import models
+
+from core.models.ticket_comment_action import TicketCommentAction
+
+
+
+class TicketCommentActionModelLink(
+    TicketCommentAction,
+):
+
+
+    @property
+    def _base_model(self):
+
+        return TicketCommentAction
+
+
+
+    class Meta:
+
+        ordering = [
+            'id'
+        ]
+
+        permissions = [
+            (
+                'import_ticketcommentactionmodellink',
+                'Can import ticket model link action comment.'
+            ),
+        ]
+
+        verbose_name = "Ticket Linked Model Action Comment"
+
+        verbose_name_plural = "Ticket Linked Model Action Comments"
+
+
+    is_create = models.BooleanField(
+        blank = False,
+        default = True,
+        help_text = 'is this an add operation?',
+        verbose_name = 'Adding link'
+    )
+
+    model_id = models.IntegerField(
+        blank = False,
+        default = 0,
+        help_text = 'ID of model',
+        null = False,
+        verbose_name = 'Model ID',
+    )
+
+    content_id = models.ForeignKey(
+        to = ContentType,
+        blank = False,
+        help_text = 'Content ID of the model',
+        null = True,
+        on_delete = models.PROTECT,
+        related_name = '+',
+        verbose_name = 'Content ID for model',
+    )
+
+
+
+    def __str__(self):
+
+        comment = f"{self.user} Linked model {self.model_id}"
+
+        if not self.is_create:
+
+            comment = f"{self.user} Unlinked model {self.model_id}"
+
+
+        return comment